Improved the sign-up process to allow connecting existing accounts, too

+++ matebot_telegram/client.py
@@ -4,7 +4,7 @@ MateBot SDK client to be used across the project
 
 import asyncio
 import logging
-from typing import Union
+from typing import Optional, Union
 
 import telegram
 
@@ -44,21 +44,14 @@ class AsyncMateBotSDKForTelegram(AsyncSDK):
         with persistence.get_new_session() as session:
             with session.begin():
                 users = session.query(persistence.TelegramUser).filter_by(telegram_id=user.id).all()
-                if len(users) == 0:
-                    session.add(persistence.TelegramUser(
-                        telegram_id=user.id,
-                        user_id=None,
-                        first_name=user.first_name,
-                        last_name=user.last_name,
-                        username=user.username
-                    ))
-                elif len(users) == 1:
+                if len(users) == 1:
                     db_user = users[0]
                     db_user.first_name = user.first_name
                     db_user.last_name = user.last_name
                     db_user.username = user.username
                     session.add(db_user)
-                else:
+                    session.commit()
+                elif len(users) > 1:
                     raise RuntimeError(f"Multiple user results for telegram ID {user.id}! Please file a bug report.")
 
     @staticmethod
@@ -85,21 +78,48 @@ class AsyncMateBotSDKForTelegram(AsyncSDK):
                     )
         raise err.AmbiguousUserSpec(f"Multiple users found for '{identifier}'. Please ensure unambiguous usernames.")
 
-    async def sign_up_new_user(self, telegram_user: telegram.User, username: str) -> _User:
-        user = await super().create_app_user(username, str(telegram_user.id), True)
-        with self.get_new_session() as session:
-            session.add(persistence.TelegramUser(
-                telegram_id=telegram_user.id,
-                first_name=telegram_user.first_name,
-                last_name=telegram_user.last_name,
-                username=telegram_user.username,
-                user_id=user.id,
-            ))
+    async def _ensure_not_existing(self, session: persistence.Session, telegram_user: telegram.User) -> Optional[_User]:
+        existing_user = session.query(persistence.TelegramUser).get(telegram_user.id)
+        if existing_user is not None:
             record = session.query(persistence.RegistrationProcess).get(telegram_user.id)
             if record is not None:
                 session.delete(record)
-            session.commit()
-        return user
+                session.commit()
+            return await self.get_user(existing_user.user_id)
+        return None
+
+    @staticmethod
+    def _handle_new_user_update(user_id: int, telegram_user: telegram.User, session: persistence.Session):
+        record = session.query(persistence.RegistrationProcess).get(telegram_user.id)
+        if record is not None:
+            session.delete(record)
+        session.commit()
+        session.add(persistence.TelegramUser(
+            telegram_id=telegram_user.id,
+            first_name=telegram_user.first_name,
+            last_name=telegram_user.last_name,
+            username=telegram_user.username,
+            user_id=user_id
+        ))
+        session.commit()
+
+    async def sign_up_as_alias(self, telegram_user: telegram.User, core_user_id: int) -> _User:
+        with self.get_new_session() as session:
+            existing_user = await self._ensure_not_existing(session, telegram_user)
+            if existing_user is not None:
+                return existing_user
+            alias = await self.create_alias(core_user_id, str(telegram_user.id), confirmed=False)
+            self._handle_new_user_update(alias.user_id, telegram_user, session)
+            return await self.get_user(alias.user_id)
+
+    async def sign_up_new_user(self, telegram_user: telegram.User, username: str) -> _User:
+        with self.get_new_session() as session:
+            existing_user = await self._ensure_not_existing(session, telegram_user)
+            if existing_user is not None:
+                return existing_user
+            user = await super().create_app_user(username, str(telegram_user.id), True)
+            self._handle_new_user_update(user.id, telegram_user, session)
+            return user
 
     async def get_core_user(self, identifier: Union[int, str, telegram.User]) -> _User:
         pretty = None
